requirements.txt
accelerate==1.10.1
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
datasets==4.5.0
dill==0.4.0
etelemetry==0.3.1
evaluate==0.4.6
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.2.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pathlib==1.0.1
pluggy==1.6.0
propcache==0.4.1
prov==2.1.1
psutil==7.2.2
puremagic==1.30
pyarrow==21.0.0
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.3.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.6.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scipy==1.13.1
simplejson==3.20.2
six==1.17.0
starlette==0.49.3
sympy==1.14.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.8.0
tqdm==4.67.3
traits==7.1.0
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
